refactor: move FIlterUI into folder
